refactor(reservation): waitingNumber - Long -> Integer 변경

+++ src/main/java/roomescape/domain/reservation/ReservationRepository.java
@@ -14,7 +14,7 @@ public interface ReservationRepository {
 
     List<Reservation> findReservations(String username);
 
-    Long countByReservationSlotId(Long reservationSlotId);
+    Integer countByReservationSlotId(Long reservationSlotId);
 
     List<Reservation> findAllByReservationIdOrder(Long reservationId);
